moved test to tests dir, misc style changes
