+++ etabotsite/etabotapp/tests/test_djangoindexhtml_mod/test_index_modified.html
@@ -1,0 +1,23 @@
+{% load staticfiles %}
+<!doctype html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <title>Welcome to ETAbot!</title>
+  <base href="/">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <link rel="icon" type="image/x-icon" href="http://www.etabot.ai/wp-content/uploads/2017/09/ETAbot-150x150.png">
+  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+  <link rel='stylesheet prefetch' href='https://cdnjs.cloudflare.com/ajax/libs/materialize/0.97.7/css/materialize.css'>
+  <link href="https://fonts.googleapis.com/css?family=Josefin+Sans|Julius+Sans+One&amp;subset=latin-ext" rel="stylesheet">
+<!-- 	<link rel='stylesheet prefetch' href='https://cdnjs.cloudflare.com/ajax/libs/materialize/0.97.7/fonts/roboto/Roboto-Bold.eot'> -->
+	<script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/2.0.0/jquery.min.js">
+    
+  </script>
+<link rel="stylesheet" href="{% static 'ng_app_js/styles.490a3be6bf3a925a56fd.css' %}"></head>
+<body>
+  <app-root><!-- Loading... --></app-root>
+<script type="text/javascript" src="{% static 'ng_app_js/runtime.a66f828dca56eeb90e02.js' %}"></script><script type="text/javascript" src="{% static 'ng_app_js/polyfills.f8faf1c1d59e26e28d42.js' %}"></script><script type="text/javascript" src="{% static 'ng_app_js/main.a30f03e4fa11708c7857.js' %}"></script></body>
+<link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" rel="stylesheet">
+<script src="https://apis.google.com/js/platform.js"></script>
+</html>
